+++ src/prefect/engine/task_runner.py
@@ -45,6 +45,7 @@ from prefect.engine.state import (
     Success,
     TimedOut,
     TriggerFailed,
+    ConditionNotMet,
 )
 from prefect.utilities.executors import (
     run_with_heartbeat,
@@ -251,6 +252,11 @@ class TaskRunner(Runner):
                     state, upstream_states=upstream_states
                 )
 
+                # check if any upstream task conditions were not met (and if we need to skip)
+                state = self.check_upstream_conditions_met(
+                    state, upstream_states=upstream_states
+                )
+
                 # if the task is mapped, process the mapped children and exit
                 if mapped:
                     state = self.run_mapped_task(
@@ -413,6 +419,41 @@ class TaskRunner(Runner):
                     )
                 )
             )
+        return state
+
+    @call_state_handlers
+    def check_upstream_conditions_met(
+        self, state: State, upstream_states: Dict[Edge, State]
+    ) -> State:
+        """
+        Checks if any of the upstream tasks have unmet conditions.
+
+        Args:
+            - state (State): the current state of this task
+            - upstream_states (Dict[Edge, State]): the upstream states
+
+        Returns:
+            - State: the state of the task after running the check
+        """
+
+        all_states = set()  # type: Set[State]
+        for edge, upstream_state in upstream_states.items():
+
+            # if the upstream state is Mapped, and this task is also mapped,
+            # we want each individual child to determine if it should
+            # skip or not based on its upstream parent in the mapping
+            if isinstance(upstream_state, Mapped) and not edge.mapped:
+                all_states.update(upstream_state.map_states)
+            else:
+                all_states.add(upstream_state)
+
+        if any(isinstance(s, ConditionNotMet) for s in all_states):
+            self.logger.debug(
+                "Task '{name}': Upstream conditions were not met; ending run.".format(
+                    name=prefect.context.get("task_full_name", self.task.name)
+                )
+            )
+            raise ENDRUN(state=Skipped(message=("Upstream task condition was not met")))
         return state
 
     @call_state_handlers