+++ tests/engine/test_state.py
@@ -14,6 +14,7 @@ from prefect.engine.state import (
     Cancelled,
     Cached,
     ClientFailed,
+    ConditionNotMet,
     Failed,
     Finished,
     Looped,
@@ -443,7 +444,7 @@ def test_children_method_on_leaf_state_returns_empty():
 
 
 def test_children_method_on_success():
-    expected = {Cached, Mapped, Skipped}
+    expected = {Cached, Mapped, Skipped, ConditionNotMet}
     assert set(Success.children()) == expected